refactor: rename mods to 'evaluators'
